# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.errorResponse:
    properties:
      code:
        type: string
      message:
        type: string
      original_error:
        type: string
    type: object
...
  models.Role:
    enum:
    - ADMIN
    - USER
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleUser
  models.User:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        $ref: '#/definitions/models.Role'
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for the study tracker backend
  title: Study Tracker API
  version: "1.0"
paths:
...
  /api/users/me:
    get:
      description: Return the authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.errorResponse'
      security:
      - token: []
      summary: Get current user
      tags:
      - users
securityDefinitions:
  token:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"